Update mgmt subcommands
This commit adds disconnects and ls subcommands and renames the file from connect to conn as the subcommands in this file deals with the connection to the Manager server of the skywire visors..
